Don't decode all images on shell activation if decodeondraw is enabled. Fixes JB#25287
Currently if the page contains a lot images in nsDocument::mImageTracker
they will be all decoded when nsDocument::SetImageLockingState(true).
This seems to be an unwanted behaviour when image.mem.decodeondraw pref
is enabled. In such case the images should be decoded just before being
painted. Since some of the images may not be even visible, we don't want to
waste memory for the decoded data.

The original fix for the problem disabled image locking completely.
Unfortunately this cused regressions on pages like jolla.com/tablet (see
JB#30305).

Signed-off-by: Raine Makelainen <raine.makelainen@jolla.com>